Separate compiled injector tests into dedicated test class
Move compile-related tests from PsrContainerTest to new CompiledPsrContainerTest class for better organization and clarity.
