+++ tests/PsrContainerTest.php
@@ -7,29 +7,14 @@ namespace NaokiTsuchiya\RayDiPsrContainer;
 use NaokiTsuchiya\RayDiPsrContainer\Attribute\Left;
 use NaokiTsuchiya\RayDiPsrContainer\Exception\ContainerException;
 use NaokiTsuchiya\RayDiPsrContainer\Exception\InvalidIdException;
-use PHPUnit\Framework\Attributes\RequiresPhp;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\NotFoundExceptionInterface;
-use Ray\Compiler\CompiledInjector;
-use Ray\Compiler\CompileInjector;
-use Ray\Compiler\Compiler;
 use Ray\Di\Injector;
-
-use const DIRECTORY_SEPARATOR;
 
 final class PsrContainerTest extends TestCase
 {
-    private const TMP_DIR = __DIR__ . DIRECTORY_SEPARATOR . 'tmp' . DIRECTORY_SEPARATOR . 'compile';
-
     private PsrContainer $injector;
-
-    public static function setUpBeforeClass(): void
-    {
-        (new ScriptDir())->init(self::TMP_DIR);
-
-        parent::setUpBeforeClass();
-    }
 
     protected function setUp(): void
     {
@@ -113,30 +98,6 @@ final class PsrContainerTest extends TestCase
         self::expectException(ContainerException::class);
 
         (new PsrContainer(new FakeExceptionThrowInjector()))->get(FakeRobotInterface::class);
-    }
-
-    #[Test]
-    #[RequiresPhp('< 8.5')]
-    public function getWithCompileInjector(): void
-    {
-        $injector = new CompileInjector(self::TMP_DIR, new FakeLazyModule());
-        $container = new PsrContainer($injector);
-        $actual = $container->get(FakeRobotInterface::class);
-
-        self::assertInstanceOf(FakeRobotInterface::class, $actual);
-        self::assertInstanceOf(FakeRobot::class, $actual);
-    }
-
-    #[Test]
-    public function getWithCompiledInjector(): void
-    {
-        (new Compiler())->compile(new FakeModule(), self::TMP_DIR);
-        $injector = new CompiledInjector(self::TMP_DIR);
-        $container = new PsrContainer($injector);
-        $actual = $container->get(FakeRobotInterface::class);
-
-        self::assertInstanceOf(FakeRobotInterface::class, $actual);
-        self::assertInstanceOf(FakeRobot::class, $actual);
     }
 
     #[Test]
